HW3 Project3 Term long double constructor -> double double
When coefficient changed to double that constructor was forgotten.

+++ hw3-proj2/src/edu/miracosta/cs113/Term.java
@@ -61,7 +61,7 @@ public class Term implements Comparable<Term>
 	 * @param exponent
 	 * @see #Term(boolean, long, String, long)
 	 */
-	public Term (long coefficient, double exponent)
+	public Term (double coefficient, double exponent)
 	{
 		this (coefficient, "x", exponent);
 	}
